ref: refactored inner eval loop (#3141)
* refactored dataloader process hook

* refactored dataloader process hook

* refactored dataloader process hook
