+++ pytorch_lightning/trainer/evaluate_loop.py
@@ -1,6 +1,7 @@
 import torch
 from pytorch_lightning.trainer.supporters import PredictionCollection
-from pytorch_lightning.core.step_result import EvalResult
+from pytorch_lightning.core.step_result import Result, EvalResult
+from pytorch_lightning.utilities.exceptions import MisconfigurationException
 
 
 class EvaluationLoop(object):
@@ -43,11 +44,38 @@ class EvaluationLoop(object):
         else:
             self.trainer.call_hook('on_validation_epoch_start', *args, **kwargs)
 
-    def evaluation_step(self, *args, **kwargs):
+    def build_args(self, test_mode, batch, batch_idx, dataloader_idx):
+        # make dataloader_idx arg in validation_step optional
+        args = [batch, batch_idx]
+
+        multiple_val_loaders = (not test_mode and len(self.trainer.val_dataloaders) > 1)
+        multiple_test_loaders = (test_mode and len(self.trainer.test_dataloaders) > 1)
+
+        if multiple_test_loaders or multiple_val_loaders:
+            args.append(dataloader_idx)
+
+        return args
+
+    def evaluation_step(self, test_mode, batch, batch_idx, dataloader_idx):
+        # configure args
+        args = self.build_args(test_mode, batch, batch_idx, dataloader_idx)
+
+        # run actual test step
         if self.testing:
-            output = self.trainer.accelerator_backend.test_step(*args, **kwargs)
+            output = self.trainer.accelerator_backend.test_step(args)
         else:
-            output = self.trainer.accelerator_backend.validation_step(*args, **kwargs)
+            output = self.trainer.accelerator_backend.validation_step(args)
+
+        # track batch size for weighted average
+        is_result_obj = isinstance(output, Result)
+        if is_result_obj:
+            output.track_batch_size(len(batch))
+
+        # allow only EvalResult when using structured results (from val_step)
+        if is_result_obj and not isinstance(output, EvalResult):
+            m = 'only EvalResults or dicts are allowed from validation_step'
+            raise MisconfigurationException(m)
+
         return output
 
     def evaluation_step_end(self, *args, **kwargs):
@@ -69,8 +97,37 @@ class EvaluationLoop(object):
         else:
             self.trainer.call_hook('on_validation_batch_end', *args, **kwargs)
 
+    def evaluation_batch_end_cleanup(self, output, batch_idx, dataloader_idx):
+        # Add step predictions to prediction collection to write later
+        if output is not None:
+            do_write_predictions = isinstance(output, Result) and self.testing
+            if do_write_predictions:
+                self.predictions.add(output.pop('predictions', None))
+
+        # track debug metrics
+        self.trainer.dev_debugger.track_eval_loss_history(self.testing, batch_idx, dataloader_idx, output)
+
     def on_evaluation_epoch_end(self, *args, **kwargs):
         if self.testing:
             self.trainer.call_hook('on_test_epoch_end', *args, **kwargs)
         else:
             self.trainer.call_hook('on_validation_epoch_end', *args, **kwargs)
+
+    def log_metrics(self, output, batch_idx):
+        if self.trainer.running_sanity_check:
+            return
+
+        if isinstance(output, EvalResult):
+            step_log_metrics = output.batch_log_metrics
+            step_pbar_metrics = output.batch_pbar_metrics
+
+            if len(step_log_metrics) > 0:
+                # make the metrics appear as a different line in the same graph
+                metrics_by_epoch = {}
+                for k, v in step_log_metrics.items():
+                    metrics_by_epoch[f'{k}/epoch_{self.trainer.current_epoch}'] = v
+
+                self.trainer.log_metrics(metrics_by_epoch, {}, step=batch_idx)
+
+            if len(step_pbar_metrics) > 0:
+                self.trainer.add_progress_bar_metrics(step_pbar_metrics)
